chore: extract api keys methods to its own file
